src/main/java/org/texastorque/toast/requirements.txt
flask
opencv-python
numpy
scipy
pupil_apriltags
pynetworktables